requirements.txt
accelerate==0.29.3
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.43.1
huggingface-hub==0.25.0
tokenizers==0.19.1
datasets==2.18.0
decord==0.6.0
deepspeed==0.16.5
diffusers==0.25.0
einops==0.6.1
ftfy==6.1.1
gradio==4.10.0
imageio==2.31.1
iopath==0.1.10
ipdb==0.13.13
joblib==1.3.1
matplotlib==3.7.1
mdtex2html==1.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
peft==0.3.0
Pillow==9.5.0
pytorchvideo==0.1.5
PyYAML==6.0
regex==2023.6.3
scipy
timm==0.9.2
tqdm==4.65.0
omegaconf==2.3.0
tensorboard==2.13.0
psutil==5.9.4
pyyaml==6.0
webdataset==0.2.48
opencv-python==4.7.0.72
sentence-transformers
bitsandbytes==0.45.4
scikit-image
visual-genome
wandb
imgaug
xformers==0.0.22.post7
spaces==0.19.4
httpx==0.27.0
safetensors==0.5.0
h5py
SentencePiece


# Core dependencies
qwen-omni-utils==0.0.3
librosa==0.11.0
ffmpeg==1.4
ffmpeg-python==0.2.0
soundfile==0.13.1
modelscope_studio==1.2.2
# git+https://github.com/huggingface/transformers@3a1ead0aabed473eafe527915eea8c197d424356
# accelerate
av

# Optional dependency
# Uncomment the following line if you need flash-attn
# flash-attn==2.7.4.post1
